pyproject.toml
[tool.poetry]
name = "arielbot"
version = "0.1.2"
description = ""
authors = ["DMC <lzxder@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
nonebot2 = {extras = ["aiohttp"], version = "^2.1.1"}
nonebot-adapter-satori = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
ariel = "arielbot.__init__:main"


[[tool.poetry.source]]
name = "ts"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true
secondary = false